+++ src/main/java/com/fitpet/server/user/application/service/UserServiceImpl.java
@@ -11,10 +11,10 @@ import com.fitpet.server.user.domain.exception.DuplicateEmailException;
 import com.fitpet.server.user.domain.exception.DuplicateNicknameException;
 import com.fitpet.server.user.domain.exception.UserNotFoundException;
 import com.fitpet.server.user.domain.repository.UserRepository;
-import com.fitpet.server.user.presentation.dto.UserCreateRequest;
 import com.fitpet.server.user.presentation.dto.UserDto;
-import com.fitpet.server.user.presentation.dto.UserInputInfoRequest;
-import com.fitpet.server.user.presentation.dto.UserUpdateRequest;
+import com.fitpet.server.user.presentation.dto.request.UserCreateRequest;
+import com.fitpet.server.user.presentation.dto.request.UserInputInfoRequest;
+import com.fitpet.server.user.presentation.dto.request.UserUpdateRequest;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
