+++ src/main/java/com/fitpet/server/user/presentation/controller/UserController.java
@@ -5,12 +5,12 @@ import com.fitpet.server.user.application.dto.GenderRankingResult;
 import com.fitpet.server.user.application.dto.RankingResult;
 import com.fitpet.server.user.application.service.UserService;
 import com.fitpet.server.user.domain.entity.Gender;
-import com.fitpet.server.user.presentation.dto.GenderRankingResponse;
-import com.fitpet.server.user.presentation.dto.RankingResponse;
-import com.fitpet.server.user.presentation.dto.UserCreateRequest;
 import com.fitpet.server.user.presentation.dto.UserDto;
-import com.fitpet.server.user.presentation.dto.UserInputInfoRequest;
-import com.fitpet.server.user.presentation.dto.UserUpdateRequest;
+import com.fitpet.server.user.presentation.dto.request.UserCreateRequest;
+import com.fitpet.server.user.presentation.dto.request.UserInputInfoRequest;
+import com.fitpet.server.user.presentation.dto.request.UserUpdateRequest;
+import com.fitpet.server.user.presentation.dto.response.GenderRankingResponse;
+import com.fitpet.server.user.presentation.dto.response.RankingResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
