+++ src/main/java/com/fitpet/server/user/presentation/controller/UserDeviceController.java
@@ -3,8 +3,8 @@ package com.fitpet.server.user.presentation.controller;
 import com.fitpet.server.shared.annotation.AuthUser;
 import com.fitpet.server.user.application.dto.DeviceTokenCommand;
 import com.fitpet.server.user.application.service.UserDeviceService;
-import com.fitpet.server.user.presentation.dto.DeviceDeleteRequestDto;
-import com.fitpet.server.user.presentation.dto.DeviceTokenRequestDto;
+import com.fitpet.server.user.presentation.dto.request.DeviceDeleteRequestDto;
+import com.fitpet.server.user.presentation.dto.request.DeviceTokenRequestDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
