+++ src/main/java/com/fitpet/server/user/presentation/dto/request/DeviceTokenRequestDto.java
@@ -1,4 +1,4 @@
-package com.fitpet.server.user.presentation.dto;
+package com.fitpet.server.user.presentation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
