+++ src/main/java/com/fitpet/server/user/presentation/dto/response/GenderRankingResponse.java
@@ -1,8 +1,8 @@
-package com.fitpet.server.user.presentation.dto;
-
-import java.util.List;
+package com.fitpet.server.user.presentation.dto.response;
 
 import com.fitpet.server.user.application.dto.GenderRankingResult;
+import com.fitpet.server.user.presentation.dto.UserRankingDto;
+import java.util.List;
 
 public record GenderRankingResponse(
         List<UserRankingDto> top10
@@ -11,8 +11,8 @@ public record GenderRankingResponse(
     public static GenderRankingResponse from(GenderRankingResult result) {
         return new GenderRankingResponse(
                 result.top10().stream()
-                    .map(UserRankingDto::from)
-                    .toList()
+                        .map(UserRankingDto::from)
+                        .toList()
         );
     }
 }
